requirements.txt
matplotlib==3.7.1
numpy==1.24.3
Pillow==9.5.0
